TINKERPOP-2245 Moved SessionExecutor/SessionFutureTask
Put SessionExecutor with all the session stuff and pushed the SessionFutureTask into SessionExecutor since it really is used from within that context.
