+++ gremlin-server/src/main/java/org/apache/tinkerpop/gremlin/server/handler/AbstractSession.java
@@ -175,6 +175,7 @@ public abstract class AbstractSession implements Session, AutoCloseable {
         return transactionManaged;
     }
 
+    @Override
     public String getSessionId() {
         return sessionId;
     }
