--- gremlin-server/src/main/java/org/apache/tinkerpop/gremlin/server/util/SessionFutureTask.java
@@ -1,44 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.tinkerpop.gremlin.server.util;
-
-import org.apache.tinkerpop.gremlin.server.handler.Session;
-
-import java.util.Optional;
-import java.util.concurrent.FutureTask;
-
-/**
- * A cancellable asynchronous operation with the added ability to get a {@link Session} instance if the
- * {@code Runnable} for the task was of that type.
- */
-public class SessionFutureTask<V> extends FutureTask<V> {
-
-    private final Session session;
-
-    public SessionFutureTask(final Runnable runnable, final  V result) {
-        super(runnable, result);
-
-        // hold an instance to the Session instance if it is of that type
-        this.session = runnable instanceof Session ? (Session) runnable : null;
-    }
-
-    public Optional<Session> getSession() {
-        return Optional.ofNullable(this.session);
-    }
-}
